driver: timer: npcx: fix possible vulnerabilities
#1 The timer reading routine is not robust enough due to an asynchronous timing issue in the chip. The workaround is implements in the assembly code to ensure it is not affected by the compiler toolchain or optimization flags. Additionally, add at least two NOPs between the LDR and CMP instructions.

#2 The ITIM driver uses an inappropriate type for the elapsed time and time ticks calculation, which could lead to data truncation with large count values. The truncated value may impact late tick calculation and settings.
For example, when passing the IN_MAX for timeout settings, the time tick will be set to 211 seconds. This causes the EC to resume from deep sleep mode earlier than expected. Additionally, the timeout setting will be configured with a large value after resumed from suspend mode, resulting in the thread not being serviced according to the expected schedule.